Merge most queries into a shared SQL file (#103)
I'm finding creating and jumping between tiny files to be quite a chore.

This leaves all the EDN for now. We probably want an option to put that inline in the SQL file, but for larger files I want a rich editor experience, without needing a [fancy Emacs mode](https://github.com/polymode/polymode).