Return error values rather than throwing (#108)
Closes https://github.com/metabase/macaw/issues/93

In Metabase we decided to clean up error handling based on values, as opposed to a mix semantically meaningful `nil`, magic keywords, and thrown exceptions.

With this change we turn (expected) exceptions into error maps. We also wrap the `tables` in a map, to avoid an awkward untagged union with the error maps.

Metabase has been made agnostic to these API changes, so upgrading will be easy.