More testing for SQL trickery (#109)
This PR improves the ergonomics for adding queries which all have the same error, and uses that to add a bunch more "skullduggery detected" fixtures.

Unfortunately I realize that a case has slipped through, and the fix is non-trivial. Needs further discussion.

Also fixed an issue where the "all the components" function was not actually catching JSQLParser exceptions.