More dynamic testing. Find a case we don't handle yet (#110)
Closes #91 
Closes #73 

This upgrades us to the latest JSQLParser, which fixes one of our fixtures.

Unfortunately it seems there is a bug in the tokens, they are giving off-by-one errors on their line, which was a real pain to discover after going down so many other rabbit holes. They also changed the scope of the Table node to include the alias, which was also quite annoying.

The new Visitor interfaces are really more Foldable / Traversibles, which is great - it means we can use them to make the completeness of our newer walkers statically verified, but I have not done that here yet.